feat: implement new procrastination strategies and corresponding tests; add documentation for collapse behaviors
