+++ src/ProcrastiN8/Services/IProcrastinationStrategy.cs
@@ -1,0 +1,17 @@
+using ProcrastiN8.Common;
+using ProcrastiN8.JustBecause;
+using ProcrastiN8.LazyTasks;
+
+namespace ProcrastiN8.Services;
+
+public interface IProcrastinationStrategy
+{
+    Task ExecuteAsync(
+        Func<Task> task,
+        TimeSpan initialDelay,
+        IExcuseProvider? excuseProvider,
+        IDelayStrategy delayStrategy,
+        IRandomProvider randomProvider,
+        ITimeProvider timeProvider,
+        CancellationToken cancellationToken);
+}
