+++ src/ProcrastiN8/Services/ProcrastinationScheduler.cs
@@ -40,100 +40,3 @@ public static class ProcrastinationScheduler
         await strategy.ExecuteAsync(task, initialDelay, excuseProvider, delayStrategy, randomProvider, timeProvider, cancellationToken);
     }
 }
-
-public interface IProcrastinationStrategy
-{
-    Task ExecuteAsync(
-        Func<Task> task,
-        TimeSpan initialDelay,
-        IExcuseProvider? excuseProvider,
-        IDelayStrategy delayStrategy,
-        IRandomProvider randomProvider,
-        ITimeProvider timeProvider,
-        CancellationToken cancellationToken);
-}
-
-public class MovingTargetStrategy : IProcrastinationStrategy
-{
-    public async Task ExecuteAsync(
-        Func<Task> task,
-        TimeSpan initialDelay,
-        IExcuseProvider? excuseProvider,
-        IDelayStrategy delayStrategy,
-        IRandomProvider randomProvider,
-        ITimeProvider timeProvider,
-        CancellationToken cancellationToken)
-    {
-        var delay = initialDelay;
-
-        while (!cancellationToken.IsCancellationRequested)
-        {
-            await delayStrategy.DelayAsync(delay, cancellationToken: cancellationToken);
-            delay = TimeSpan.FromMilliseconds(delay.TotalMilliseconds * (1 + randomProvider.GetDouble() * 0.25));
-
-            if (excuseProvider != null)
-            {
-                await excuseProvider.GetExcuseAsync();
-            }
-
-            if (delay.TotalMilliseconds > TimeSpan.FromHours(1).TotalMilliseconds)
-            {
-                break;
-            }
-        }
-
-        await task();
-    }
-}
-
-public class InfiniteEstimationStrategy : IProcrastinationStrategy
-{
-    public async Task ExecuteAsync(
-        Func<Task> task,
-        TimeSpan initialDelay,
-        IExcuseProvider? excuseProvider,
-        IDelayStrategy delayStrategy,
-        IRandomProvider randomProvider,
-        ITimeProvider timeProvider,
-        CancellationToken cancellationToken)
-    {
-        while (!cancellationToken.IsCancellationRequested)
-        {
-            if (excuseProvider != null)
-            {
-                await excuseProvider.GetExcuseAsync();
-            }
-            await delayStrategy.DelayAsync(TimeSpan.FromMinutes(5), cancellationToken: cancellationToken);
-        }
-    }
-}
-
-public class WeekendFallbackStrategy : IProcrastinationStrategy
-{
-    public async Task ExecuteAsync(
-        Func<Task> task,
-        TimeSpan initialDelay,
-        IExcuseProvider? excuseProvider,
-        IDelayStrategy delayStrategy,
-        IRandomProvider randomProvider,
-        ITimeProvider timeProvider,
-        CancellationToken cancellationToken)
-    {
-        while (!cancellationToken.IsCancellationRequested)
-        {
-            var now = timeProvider.GetUtcNow();
-            if (now.DayOfWeek == DayOfWeek.Saturday && now.Hour >= 15 ||
-                now - timeProvider.GetUtcNow() >= TimeSpan.FromHours(72))
-            {
-                await task();
-                return;
-            }
-
-            if (excuseProvider != null)
-            {
-                await excuseProvider.GetExcuseAsync();
-            }
-            await delayStrategy.DelayAsync(TimeSpan.FromHours(1), cancellationToken: cancellationToken);
-        }
-    }
-}
